# backend/cmd/order-service/docs/swagger.yaml
basePath: /
definitions:
...
  domain.Product:
    properties:
      category:
        type: string
      created_at:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        description: unit price in cents
        type: integer
      sku:
        type: string
      status:
        $ref: '#/definitions/domain.ProductStatus'
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  domain.ProductStatus:
    enum:
    - active
    - inactive
    - out_of_stock
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusOutOfStock
...
  domain.Wallet:
    properties:
      balance:
        description: cents
        type: integer
      created_at:
        type: string
      currency:
        type: string
      id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handler.CreateProductRequest:
    properties:
      category:
        example: electronics
        type: string
      currency:
        example: USD
        type: string
      description:
        example: Noise-cancelling over-ear headphones
        type: string
      name:
        example: Wireless Headphones
        type: string
      price:
        description: cents
        example: 9999
        type: integer
      sku:
        example: HDPH-001
        type: string
      stock:
        example: 100
        minimum: 0
        type: integer
    required:
    - name
    - price
    - sku
    type: object
  handler.CreateWalletRequest:
    properties:
      currency:
        example: USD
        type: string
    required:
    - currency
    type: object
...
  handler.ListProductsResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      products:
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      total:
        example: 42
        type: integer
    type: object
...
  handler.ProductErrorResponse:
    properties:
      error:
        example: product not found
        type: string
    type: object
...
  handler.UpdateProductRequest:
    properties:
      category:
        example: electronics
        type: string
      description:
        example: Updated description
        type: string
      name:
        example: Wireless Headphones Pro
        type: string
      price:
        example: 11999
        type: integer
      stock:
        example: 200
        type: integer
    type: object
...
  handler.WalletErrorResponse:
    properties:
      error:
        example: wallet not found
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: Manages order creation, retrieval, and status updates driven by payment
    events.
  title: Order Service
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      description: Returns paginated products. Filter by category, status, or a comma-separated
        list of IDs (used by order service for batch price lookup).
      parameters:
      - description: Comma-separated product IDs for batch lookup
        in: query
        name: ids
        type: string
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: 'Filter by status: active, inactive, out_of_stock'
        in: query
        name: status
        type: string
      - default: 20
        description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListProductsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      summary: List products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Adds a new product to the catalog with an authoritative price and
        initial stock.
      parameters:
      - description: Product details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
        "409":
          description: SKU already exists
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      description: Soft-deletes a product by setting its status to inactive. It remains
        in the DB.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Deactivate product
      tags:
      - products
    get:
      description: Returns product details including current price and stock level.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      summary: Get product
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: Partially updates a product. Only provided fields are changed.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ProductErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - products
  /api/v1/wallet:
    get:
      description: Returns the authenticated user's wallet balance and details.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Wallet'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
      security:
      - BearerAuth: []
      summary: Get wallet
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Creates a new wallet for the authenticated user. Each user can
        have only one wallet.
      parameters:
      - description: Wallet details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateWalletRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Wallet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.WalletErrorResponse'
      security:
      - BearerAuth: []
      summary: Create wallet
      tags:
      - wallet
...
securityDefinitions:
  BearerAuth:
    description: JWT token issued by the API Gateway
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"